Deleting 'preview' models and updating the db
This commit deletes the old models, 'Events' and 'Users', since those
were only there to explain the code initially. As we progress, we will
now be using the production models. This commit is the start of that.
