+++ code/restful-api/.vs/config/applicationhost.config
@@ -152,6 +152,7 @@
         </application>
         <bindings>
           <binding protocol="http" bindingInformation=":8080:localhost" />
+          <binding protocol="http" bindingInformation=":8080:192.168.25.4" />
         </bindings>
       </site>
       <site name="restful-api" id="2">
