--- code/restful-api/restful-api/Models/Event.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace RestfulApi.Models
-{
-    public class Event
-    {
-        public long Id { get; set; }
-        public String Name { get; set; }
-        public String Description { get; set; }
-        public long UserId { get; set; }
-        public User User { get; set; }
-    }
-}
