--- code/restful-api/restful-api/Models/User.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace RestfulApi.Models
-{
-    public class User
-    {
-        public long Id { get; set; }
-        public String Name { get; set; }
-        public String Email { get; set; }
-    }
-}
