+++ code/restful-api/restful-api/Models/Usuario.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace RestfulApi.Models
+{
+    public class Usuario 
+    {
+        public long Id { get; set; }
+        public string Nome { get; set; }
+        public string Email { get; set; }
+        public string Senha { get; set; }
+        public string Cpf { get; set; }
+    }
+}
